Check type assertion when checking kid
When the supplied token is too old and google as already changed it's
public keys, the kid supplied will not match anything and the type
assertion will create a panic due to missing key in the map. Make sure
we don't panic and the application isn't restarted which would cause the
actual request to proceed however post starting the application again on
heroku.

Signed-off-by: Pranay Kumar Srivastava <pranjas@gmail.com>